fix: split 'ide' into 'browse' and 'runtime' to avoid clash between orocos.rb and runkit
