+++ bin/syskit
@@ -15,12 +15,11 @@ if ENV["SYSKIT_USE_ROCK_BUNDLES"] != "0"
     Roby.app.public_logs = false
 end
 
-require "syskit/cli/main"
-
 # Transform 'syskit help <mode>' to 'syskit <mode> --help' if mode is not
 # handled by thor
 if ARGV[0] == "help"
     if (command_name = ARGV[1])
+        require "syskit/cli/main"
         unless Syskit::CLI::Main.all_commands.key?(command_name)
             ARGV.shift
             ARGV << "--help"
@@ -30,31 +29,42 @@ if ARGV[0] == "help"
 # by thor
 elsif ["-h", "--help"].include?(ARGV[1])
     STDERR.puts "WARN: syskit <mode> --help is deprecated, use syskit help <mode> instead"
+    require "syskit/cli/main"
     if Syskit::CLI::Main.all_commands.key?(ARGV[0])
         ARGV.unshift "help"
     end
 end
 
-if ARGV.first
+ORIGINAL_ARGV = ARGV.dup
+
+SYSKIT_MODES = %w[ide runtime browse process_server log_runtime_archive].freeze
+ROBY_MODES = %w[run shell test gen quit restart].freeze
+
+if SYSKIT_MODES.include?(ARGV.first)
+    require "syskit/scripts/#{ARGV.shift}"
+    exit(0)
+elsif ROBY_MODES.include?(ARGV.first)
+    require "roby/app/scripts/#{ARGV.shift}"
+    exit(0)
+elsif ARGV.first
     command_name = ARGV.first.tr("-", "_")
     global_help = (command_name == "help") && (ARGV.size == 1)
-    if !global_help && Syskit::CLI::Main.all_commands.key?(command_name)
-        begin
-            Syskit::CLI::Main.start(ARGV)
-            exit 0
-        rescue Roby::CLI::CLIException => e
-            STDERR.puts Roby.color(e.message, :red)
-            exit 1
+    unless global_help
+        require "syskit/cli/main"
+        if Syskit::CLI::Main.all_commands.key?(command_name)
+            begin
+                Syskit::CLI::Main.start(ARGV)
+                exit 0
+            rescue Roby::CLI::CLIException => e
+                STDERR.puts Roby.color(e.message, :red)
+                exit 1
+            end
         end
     end
 end
 
-ORIGINAL_ARGV = ARGV.dup
-mode = ARGV.shift
-
-SYSKIT_MODES = %w[ide process_server log_runtime_archive].freeze
-ROBY_MODES = %w[run shell test gen quit restart].freeze
-if [nil, "--help", "-h", "help"].include?(mode)
+if [nil, "--help", "-h", "help"].include?(ARGV.first)
+    require "syskit/cli/main"
     thor_modes = Syskit::CLI::Main
                  .all_commands
                  .find_all { |_name, command| !command.hidden? }
@@ -62,21 +72,16 @@ if [nil, "--help", "-h", "help"].include?(mode)
     all_modes = SYSKIT_MODES | ROBY_MODES | thor_modes
     puts "usage: syskit [#{all_modes.sort.join('|')}] <mode-arguments>"
     puts "Run 'syskit help <mode>' for more information"
-else
-    if SYSKIT_MODES.include?(mode)
-        require "syskit/scripts/#{mode}"
-    elsif ROBY_MODES.include?(mode)
-        require "roby/app/scripts/#{mode}"
-    end
+    return
+end
 
+begin
+    require "syskit/scripts/#{mode}"
+rescue LoadError => e
     begin
-        require "syskit/scripts/#{mode}"
-    rescue LoadError => e
-        begin
-            require "roby/app/scripts/#{mode}"
-        rescue LoadError
-            STDERR.puts "unknown mode '#{mode}'"
-            exit(1)
-        end
+        require "roby/app/scripts/#{mode}"
+    rescue LoadError
+        STDERR.puts "unknown mode '#{mode}'"
+        exit(1)
     end
 end
