+++ lib/syskit/scripts/browse.rb
@@ -1,0 +1,62 @@
+# frozen_string_literal: true
+
+require "roby"
+require "syskit/gui/ide"
+require "syskit/scripts/common"
+
+load_all = false
+test_mode = false
+parser = OptionParser.new do |opt|
+    opt.banner = <<~BANNER_TEXT
+        Usage: ide [file] [options]
+        Loads the models from this bundle and allows to browse them. If a file is given, only this file is loaded.
+    BANNER_TEXT
+
+    opt.on "--all", "-a", "Load all models from all active bundles instead of only the ones from the current" do
+        load_all = true
+    end
+
+    opt.on "-t", "--test", "Start with tests already running" do
+        test_mode = true
+    end
+end
+
+# We must require_app_dir before we call common_options, or the list of robots
+# won't be the right one. Hand-detect --runtime-only and act accordingly
+Roby.app.require_app_dir
+
+Syskit::Scripts.common_options(parser, true)
+remaining = parser.parse(ARGV)
+
+# We don't need the process server, win some startup time
+Roby.app.using "syskit"
+Syskit.conf.only_load_models = true
+Syskit.conf.disables_local_process_server = true
+Roby.app.ignore_all_load_errors = true
+Roby.app.development_mode = false
+
+direct_files, model_names = remaining.partition do |arg|
+    File.file?(arg)
+end
+# Load all task libraries if we don't get a file to require
+Roby.app.auto_load_all = load_all
+Roby.app.auto_load_models = false
+Roby.app.additional_model_files.concat(direct_files)
+
+Syskit::Scripts.run do
+    qapp = Qt::Application.new(ARGV)
+
+    main = Syskit::GUI::Browse.new(
+        robot_name: Roby.app.robot_name,
+        tests: test_mode,
+        host: options[:host], port: options[:port]
+    )
+    main.window_title = "Syskit #{Roby.app.app_name} #{Roby.app.robot_name}"
+    main.restore_from_settings
+    main.show
+
+    qapp.exec
+
+    main.save_to_settings
+    main.settings.sync
+end
